backend/requirements.txt

fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
rdkit>=2022.9.5
supabase>=2.0.0
openai>=0.28.0
reportlab>=4.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.3.0
mypy>=1.7.0
types-requests>=2.31.0
